requirements.txt
requests
transformers
nltk
rouge-score
python-dotenv
sentencepiece
torch
